// enums/applicants/applicant-onboarding-checklist.enum.ts
export enum ApplicantOnBoardingChecklist {
  EMPLOYMENT_APPLICATION = "EMPLOYMENT_APPLICATION",
  DRIVER_AUTHORIZATION = "DRIVER_AUTHORIZATION",
  RESUME_OPTIONAL = "RESUME_OPTIONAL",
  OFFER_LETTER_AND_ACCEPTANCE = "OFFER_LETTER_AND_ACCEPTANCE",
  DRIVER_LICENSE = "DRIVER_LICENSE",
  MOTOR_VEHICLE_RECORD_MVR = "MOTOR_VEHICLE_RECORD_MVR",
  DRIVER_ROAD_TEST_CERTIFICATE_OR_EQUIVALENT = "DRIVER_ROAD_TEST_CERTIFICATE_OR_EQUIVALENT",
  DRUG_AND_ALCOHOL_CLEARINGHOUSE_QUERY = "DRUG_AND_ALCOHOL_CLEARINGHOUSE_QUERY",
  DRUG_AND_ALCOHOL_TEST_CHAIN_OF_CUSTODY = "DRUG_AND_ALCOHOL_TEST_CHAIN_OF_CUSTODY",
  MEDICAL_EXAMINER_CERTIFICATE_MEDICAL_CARD = "MEDICAL_EXAMINER_CERTIFICATE_MEDICAL_CARD",
  MEDICAL_EXAMINATION_REPORT_LONG_FORM = "MEDICAL_EXAMINATION_REPORT_LONG_FORM",
  SAFETY_PERFORMANCE_HISTORY = "SAFETY_PERFORMANCE_HISTORY",
  RECORD_OF_ATTEMPTS_SAFETY_PERFORMANCE_HISTORY = "RECORD_OF_ATTEMPTS_SAFETY_PERFORMANCE_HISTORY",
  PSP_RECORD = "PSP_RECORD",
  PRE_EMPLOYMENT_DRUG_TEST = "PRE_EMPLOYMENT_DRUG_TEST",
  PROOF_OF_INSURANCE = "PROOF_OF_INSURANCE",
  VEHICLE_TITLE = "VEHICLE_TITLE",
  EQUIPMENT_LEASE = "EQUIPMENT_LEASE",
  SIGNED_HOURS_OF_SERVICE_RECORD_STATEMENT = "SIGNED_HOURS_OF_SERVICE_RECORD_STATEMENT",
  TWIC_OPTIONAL = "TWIC_OPTIONAL",
}